+++ src/windows/AzPin.Windows.Core/Models/AppRunningState.cs
@@ -1,0 +1,11 @@
+namespace AzPin.Windows.Models;
+
+public enum AppRunningState
+{
+    Unknown,
+    Running,
+    Stopped,
+    Starting,
+    Stopping,
+    Restarting
+}